Patch VM config wrt. architecture.
