+++ process/factory/shard/vmConfigPatching_darwin_arm64.go.go
@@ -1,0 +1,8 @@
+//go:build darwin && arm64
+
+package shard
+
+import "github.com/multiversx/mx-chain-go/config"
+
+func patchVirtualMachineConfigGivenArchitecture(config *config.VirtualMachineConfig) {
+}
