+++ process/factory/shard/vmContainerFactory.go
@@ -87,6 +87,8 @@ func NewVMContainerFactory(args ArgVMContainerFactory) (*vmContainerFactory, err
 		return nil, process.ErrNilHasher
 	}
 
+	patchVirtualMachineConfigGivenArchitecture(&args.Config)
+
 	cryptoHook := hooks.NewVMCryptoHook()
 
 	vmf := &vmContainerFactory{
@@ -277,6 +279,20 @@ func (vmf *vmContainerFactory) getMatchingVersion(epoch uint32) config.WasmVMVer
 	}
 
 	return matchingVersion
+}
+
+func (vmf *vmContainerFactory) createInProcessWasmVMByVersion(version config.WasmVMVersionByEpoch) (vmcommon.VMExecutionHandler, error) {
+	logVMContainerFactory.Debug("createInProcessWasmVMByVersion !(darwin && arm64)", "version", version)
+	switch version.Version {
+	case "v1.2":
+		return vmf.createInProcessWasmVMV12()
+	case "v1.3":
+		return vmf.createInProcessWasmVMV13()
+	case "v1.4":
+		return vmf.createInProcessWasmVMV14()
+	default:
+		return vmf.createInProcessWasmVMV15()
+	}
 }
 
 func (vmf *vmContainerFactory) createInProcessWasmVMV12() (vmcommon.VMExecutionHandler, error) {
